Moved method generateMetaData from ApiInfo to EurekaMetadataParser
