+++ common-service-core/src/main/java/com/ca/mfaas/config/ApiInfo.java
@@ -9,15 +9,10 @@
  */
 package com.ca.mfaas.config;
 
-import com.ca.mfaas.utils.UrlUtils;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.util.HashMap;
-import java.util.Map;
-
-import static com.ca.mfaas.constants.EurekaMetadataDefinition.*;
 
 /**
  * Represents one API provided by a service
@@ -33,42 +28,4 @@ public class ApiInfo {
     private String version;
     private String swaggerUrl;
     private String documentationUrl;
-
-    /**
-     * Generate Eureka metadata for ApiInfo configuration
-     *
-     * @param serviceId the identifier of a service which ApiInfo configuration belongs
-     * @return the generated Eureka metadata
-     */
-    //TODO: move to EurekaMetadataParser
-    public Map<String, String> generateMetadata(String serviceId) {
-        Map<String, String> metadata = new HashMap<>();
-        String encodedGatewayUrl = UrlUtils.getEncodedUrl(gatewayUrl);
-
-        if (gatewayUrl != null) {
-            metadata.put(String.format("%s.%s.%s", API_INFO, encodedGatewayUrl, API_INFO_GATEWAY_URL), gatewayUrl);
-        }
-
-        if (version != null) {
-            metadata.put(String.format("%s.%s.%s", API_INFO, encodedGatewayUrl, API_INFO_VERSION), version);
-        }
-
-        if (swaggerUrl != null) {
-            UrlUtils.validateUrl(swaggerUrl,
-                () -> String.format("The Swagger URL \"%s\" for service %s is not valid", swaggerUrl, serviceId)
-            );
-
-            metadata.put(String.format("%s.%s.%s", API_INFO, encodedGatewayUrl, API_INFO_SWAGGER_URL), swaggerUrl);
-        }
-
-        if (documentationUrl != null) {
-            UrlUtils.validateUrl(documentationUrl,
-                () -> String.format("The documentation URL \"%s\" for service %s is not valid", documentationUrl, serviceId)
-            );
-
-            metadata.put(String.format("%s.%s.%s", API_INFO, encodedGatewayUrl, API_INFO_DOCUMENTATION_URL), documentationUrl);
-        }
-
-        return metadata;
-    }
 }
