+++ common-service-core/src/test/java/com/ca/mfaas/eurekaservice/client/util/EurekaMetadataParserTest.java
@@ -12,8 +12,11 @@ package com.ca.mfaas.eurekaservice.client.util;
 import com.ca.mfaas.config.ApiInfo;
 import com.ca.mfaas.product.routing.RoutedService;
 import com.ca.mfaas.product.routing.RoutedServices;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
 
+import java.security.InvalidParameterException;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -22,11 +25,14 @@ import java.util.Map;
 import static com.ca.mfaas.constants.EurekaMetadataDefinition.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 
 public class EurekaMetadataParserTest {
 
     private final EurekaMetadataParser eurekaMetadataParser = new EurekaMetadataParser();
+
+    @Rule
+    public ExpectedException exceptionRule = ExpectedException.none();
 
     @Test
     public void testParseApiInfo() {
@@ -103,4 +109,82 @@ public class EurekaMetadataParserTest {
         assertEquals("List route size is different", 3, actualRoutes.size());
         assertThat(actualRoutes, containsInAnyOrder(expectedListRoute.toArray()));
     }
+    @Test
+    public void generateFullMetadata() {
+        String serviceId = "test service";
+        String gatewayUrl = "api/v1";
+        String version = "1.0.0";
+        String swaggerUrl = "https://service/api-doc";
+        String documentationUrl = "https://www.zowe.org";
+        String metadataPrefix = API_INFO + ".api-v1.";
+
+        ApiInfo apiInfo = new ApiInfo("org.zowe", gatewayUrl, version, swaggerUrl, documentationUrl);
+        Map<String, String> metadata = eurekaMetadataParser.generateMetadata(serviceId, apiInfo);
+
+        String metaVersion = metadata.get(metadataPrefix + API_INFO_VERSION);
+        assertNotNull(metaVersion);
+        assertEquals(version, metaVersion);
+
+        String metaGatewayUrl = metadata.get(metadataPrefix + API_INFO_GATEWAY_URL);
+        assertNotNull(metaGatewayUrl);
+        assertEquals(gatewayUrl, metaGatewayUrl);
+
+        String metaSwaggerUrl = metadata.get(metadataPrefix + API_INFO_SWAGGER_URL);
+        assertNotNull(metaSwaggerUrl);
+        assertEquals(swaggerUrl, metaSwaggerUrl);
+
+        String metaDocumentationUrl = metadata.get(metadataPrefix + API_INFO_DOCUMENTATION_URL);
+        assertNotNull(metaDocumentationUrl);
+        assertEquals(documentationUrl, metaDocumentationUrl);
+    }
+
+    @Test
+    public void generateMetadataWithNoGatewayUrl() {
+        String serviceId = "test service";
+        String version = "1.0.0";
+
+        ApiInfo apiInfo = new ApiInfo(null, null, version, null, null);
+        Map<String, String> metadata = eurekaMetadataParser.generateMetadata(serviceId, apiInfo);
+
+        assertEquals(1, metadata.size());
+        assertTrue(metadata.toString().contains(version));
+    }
+
+    @Test
+    public void generateNoMetadata() {
+        String serviceId = "test service";
+
+        ApiInfo apiInfo = new ApiInfo();
+        Map<String, String> metadata = eurekaMetadataParser.generateMetadata(serviceId, apiInfo);
+
+        assertEquals(0, metadata.size());
+    }
+
+    @Test
+    public void generateMetadataWithIncorrectSwaggerUrl() {
+        String serviceId = "test service";
+        String gatewayUrl = "api/v1";
+        String swaggerUrl = "www.badAddress";
+
+        exceptionRule.expect(InvalidParameterException.class);
+        exceptionRule.expectMessage("The Swagger URL \"" + swaggerUrl + "\" for service " + serviceId + " is not valid: no protocol: " + swaggerUrl);
+
+        ApiInfo apiInfo = new ApiInfo(null, gatewayUrl, null, swaggerUrl, null);
+        eurekaMetadataParser.generateMetadata(serviceId, apiInfo);
+    }
+
+
+    @Test
+    public void generateMetadataWithIncorrectDocumentationUrl() {
+        String serviceId = "test service";
+        String gatewayUrl = "api/v1";
+        String documentationUrl = "www.badAddress";
+
+        exceptionRule.expect(InvalidParameterException.class);
+        exceptionRule.expectMessage("The documentation URL \"" + documentationUrl + "\" for service " + serviceId + " is not valid: no protocol: " + documentationUrl);
+
+        ApiInfo apiInfo = new ApiInfo(null, gatewayUrl, null, null, documentationUrl);
+        eurekaMetadataParser.generateMetadata(serviceId, apiInfo);
+    }
+
 }
