+++ discovery-service/src/main/java/com/ca/mfaas/discovery/staticdef/ServiceDefinitionProcessor.java
@@ -10,6 +10,7 @@
 package com.ca.mfaas.discovery.staticdef;
 
 import com.ca.mfaas.config.ApiInfo;
+import com.ca.mfaas.eurekaservice.client.util.EurekaMetadataParser;
 import com.ca.mfaas.utils.UrlUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
@@ -286,7 +287,7 @@ public class ServiceDefinitionProcessor {
 
             if (service.getApiInfo() != null) {
                 for (ApiInfo apiInfo : service.getApiInfo()) {
-                    mt.putAll(apiInfo.generateMetadata(service.getServiceId()));
+                    mt.putAll(EurekaMetadataParser.generateMetadata(service.getServiceId(), apiInfo));
                 }
             }
         }
